Fix video interlacing issue with vision
If the next beans changes and a new move is recognized on a frame which is in
the middle of interlacing, the next beans could get recognized incorrectly.
This could cause the next move to be incorrectly recognized. This fixes the
problem by allowing one extra frame for the next beans to settle. The next move
is still triggered on the first change of next beans though, so in the future
I'll consider waiting until it's been changed for two frames until returning
`new_move`==True.